refactors code to make it more modular
